requirements.txt
Django==4.1.4
numpy==1.23.5
Pillow==9.4.0